requirements.txt
openai-whisper
torch
torchaudio
ffmpeg-python